@@ -12,7 +12,7 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "eslint": "1.0.x",
+    "eslint": "1.2.x",
     "eslint-plugin-react": "3.2.x",
     "express": "4.x"
   },
