@@ -84,9 +84,9 @@ add_task(function* test_nouser() {
   Assert.ok(!panelUIFooter.hasAttribute("fxastatus"), "no fxsstatus when signed out");
   Assert.ok(!panelUIFooter.hasAttribute("fxaprofileimage"), "no fxaprofileimage when signed out");
 
-  let promiseOpen = promiseTabOpen("about:accounts?entrypoint=menupanel");
+  let promisePreferencesOpened = promiseObserver("test:browser_fxaccounts:openPreferences");
   panelUIStatus.click();
-  yield promiseOpen;
+  yield promisePreferencesOpened;
 });
 
 /*
