@@ -532,7 +532,17 @@ function openPreferences(paneID, extraArgs)
   if (getBoolPref("browser.preferences.inContent")) {
     let win = Services.wm.getMostRecentWindow("navigator:browser");
     let friendlyCategoryName = internalPrefCategoryNameToFriendlyName(paneID);
-    let preferencesURL = "about:preferences" +
+    let params;
+    if (extraArgs && extraArgs["urlParams"]) {
+      params = new URLSearchParams();
+      let urlParams = extraArgs["urlParams"];
+      for (let name in urlParams) {
+        if (urlParams[name] !== undefined) {
+          params.set(name, urlParams[name]);
+        }
+      }
+    }
+    let preferencesURL = "about:preferences" + (params ? "?" + params : "") +
                          (friendlyCategoryName ? "#" + friendlyCategoryName : "");
     let newLoad = true;
     let browser = null;
